WIP - fundamental math operators dont work for arrays of FixedPoint because SSE not correctly implemented for them

+++ libs/math/include/math/shapeless_array.hpp
@@ -627,7 +627,7 @@ public:
           vb = -vb;
         }
         double M = std::max(va, vb);
-        std::cout << this->At(i) << " " << other[i] << " "
+        std::cout << static_cast<double>(this->At(i)) << " " << static_cast<double>(other[i]) << " "
                   << ((vA < std::max(atol, M * rtol)) ? " " : "*") << std::endl;
       }
     }
